Admin

+++ components/SiteMap/controllers/class.Delete.php
@@ -23,11 +23,15 @@
 			
 			try {
 				if (empty($_POST['page']['id']))
-					throw new \forge\HttpException(_('The ID must not be empty'), \forge\HttpException::HTTP_BAD_REQUEST);
+					throw new \forge\HttpException('The ID must not be empty', \forge\HttpException::HTTP_BAD_REQUEST);
 				
 				\forge\components\Databases::DB()->beginTransaction();
 				
 				$entry = new \forge\components\SiteMap\db\Page($_POST['page']['id']);
+				
+				if (!$entry->allowRemove)
+					throw new \forge\HttpException('The page is not allowed to be removed', \forge\HttpException::HTTP_FORBIDDEN);
+				
 				(new $entry->page_type)->delete($entry->getID());
 				$entry->delete();
